+++ skyfield/framelib.py
@@ -49,11 +49,18 @@ class ICRS(object):
     def rotation_at(t):
         return _identity
 
-def build_ecliptic_matrix(t):
-    # Build the matrix to rotate an ICRF vector into ecliptic coordinates.
-    _, d_eps = t._nutation_angles_radians
-    true_obliquity = t._mean_obliquity_radians + d_eps
-    return mxm(rot_x(- true_obliquity), t.M)
+class mean_equator_and_equinox_of_date(object):
+    """The coordinate frame of Earth’s mean equator and equinox.
+
+    This frame is used for measuring right ascension and declination.
+    It tracks the Earth’s ‘mean’ equator and equinox which shift slowly
+    across the sky due to precession, but ignores the smaller effects of
+    nutation.
+
+    """
+    @staticmethod
+    def rotation_at(t):
+        return t.P
 
 class true_equator_and_equinox_of_date(object):
     """The dynamical frame of Earth’s true equator and true equinox of date.
@@ -79,8 +86,6 @@ class true_equator_and_equinox_of_date(object):
     @staticmethod
     def rotation_at(t):
         return t.M
-
-true_equator_and_equinox_of_date = true_equator_and_equinox_of_date()
 
 _itrs_angvel_matrix = array((
     (0.0, DAY_S * ANGVEL, 0.0),
@@ -139,6 +144,12 @@ class itrs(object):
         return _itrs_angvel_matrix
 
 itrs = itrs()
+
+def build_ecliptic_matrix(t):
+    # Build the matrix to rotate an ICRF vector into ecliptic coordinates.
+    _, d_eps = t._nutation_angles_radians
+    true_obliquity = t._mean_obliquity_radians + d_eps
+    return mxm(rot_x(- true_obliquity), t.M)
 
 class ecliptic_frame(object):
     """Reference frame of the true ecliptic and equinox of date."""